requirements.txt
networkx==1.11
pytorch
scikit-learn
scipy
numpy